pyproject.toml
[tool.poetry]
name = "sockio-echo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-socketio = "^5.11.0"
eventlet = "^0.35.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
